--- src/main/kotlin/io/acari/n7/config/ui/NormanyFormNew.kt
@@ -1,31 +1,0 @@
-package io.acari.n7.config.ui
-
-import com.intellij.openapi.options.SearchableConfigurable
-import javax.swing.JComponent
-
-class NormanyFormNew : SearchableConfigurable {
-  companion object {
-    const val ID = "SSV Normandy Settings"
-  }
-
-  override fun isModified(): Boolean {
-    TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-  }
-
-  override fun getId(): String {
-    TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-  }
-
-  override fun getDisplayName(): String {
-    TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-  }
-
-  override fun apply() {
-    TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-  }
-
-  override fun createComponent(): JComponent? {
-    TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
-  }
-
-}
